Factor out checks
