+++ test-table-validator.sh
@@ -9,6 +9,6 @@ for FIXTURE in table-examples/*; do
   ACTUAL="/tmp/actual-output.txt"
   CMD="python src/ingest_validation_tools/table_validator.py --fixture $FIXTURE"
   echo "Running: $CMD"
-  $CMD > $ACTUAL
+  PYTHONPATH="src:$PYTHONPATH" $CMD > $ACTUAL
   diff $EXPECTED $ACTUAL || die "To fix: $CMD > $EXPECTED"
 done
